router: accel-pptp: fix some warnings from code analyzer

+++ release/src-rt-6.x.4708/router/accel-pptp/src/dirutil.c
@@ -37,7 +37,7 @@ char *dirnamex(char *pathname)
  * stripslash("/") returns "/";
  */
 char *stripslash(char *pathname) {
-    int len = strlen(pathname);
+    size_t len = strlen(pathname);
     while (len > 1 && pathname[len - 1] == '/')
         pathname[--len] = '\0';
     return pathname;
